+++ lib/ui/fish/fish_details_screen.dart
@@ -5,6 +5,7 @@ import 'package:flutter/material.dart';
 import 'package:provider/provider.dart';
 
 import 'package:acnh_helper/hemisphere.dart';
+import 'package:acnh_helper/item_type.dart';
 import 'package:acnh_helper/model/fish.dart';
 import 'package:acnh_helper/provider/fishes_provider.dart';
 import 'package:acnh_helper/ui/common/app_colors.dart';
@@ -14,6 +15,9 @@ import 'package:acnh_helper/ui/common/times_available.dart';
 import 'package:acnh_helper/utils.dart';
 
 class FishDetailsScreen extends BaseDetailsScreen<FishesProvider, Fish> {
+  @override
+  String get itemType => ItemType.fishes;
+
   FishDetailsScreen({int itemId}): super(itemId: itemId);
 
   @override