+++ lib/ui/villager/villager_details_screen.dart
@@ -4,6 +4,7 @@ import 'package:flutter/material.dart';
 
 import 'package:provider/provider.dart';
 
+import 'package:acnh_helper/item_type.dart';
 import 'package:acnh_helper/model/villager.dart';
 import 'package:acnh_helper/provider/villagers_provider.dart';
 import 'package:acnh_helper/ui/common/app_colors.dart';
@@ -11,6 +12,9 @@ import 'package:acnh_helper/ui/common/base_details_screen.dart';
 import 'package:acnh_helper/utils.dart';
 
 class VillagerDetailsScreen extends BaseDetailsScreen<VillagersProvider, Villager> {
+  @override
+  String get itemType => ItemType.villagers;
+
   VillagerDetailsScreen({int itemId}): super(itemId: itemId);
 
   @override